RocksDbStore usable from Blockchain Toolkit
