+++ src/RocksDBStore/Plugins/Storage/RocksDBStorageProvider.cs
@@ -1,0 +1,19 @@
+using Neo.Persistence;
+
+namespace Neo.Plugins.Storage
+{
+    public class RocksDBStorageProvider : Plugin, IStorageProvider
+    {
+        public override string Description => "Uses RocksDBStore to store the blockchain data";
+
+        /// <summary>
+        /// Get store
+        /// </summary>
+        /// <returns>RocksDbStore</returns>
+        public IStore GetStore(string path)
+        {
+            path = string.Format(path, ProtocolSettings.Default.Magic.ToString("X8"));
+            return RocksDbStore.Open(path);
+        }
+    }
+}
